# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Acknowledged": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "Acknowledged",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearnerAccountUpdate": {
        "additionalProperties": false,
        "properties": {
          "alias": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Alias",
            "type": "string"
          },
          "password": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "required": [
          "alias"
        ],
        "title": "LearnerAccountUpdate",
        "type": "object"
      },
      "LearnerDevicePublic": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "expires_at"
        ],
        "title": "LearnerDevicePublic",
        "type": "object"
      },
...
      "LearnerInput": {
        "additionalProperties": false,
        "properties": {
          "alias": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Alias",
            "type": "string"
          },
          "eligibility": {
            "default": "unknown",
            "enum": [
              "adult",
              "minor",
              "unknown"
            ],
            "title": "Eligibility",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "alias",
          "password"
        ],
        "title": "LearnerInput",
        "type": "object"
      },
      "LearnerPublic": {
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "eligibility": {
            "enum": [
              "adult",
              "minor",
              "unknown"
            ],
            "title": "Eligibility",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "has_password": {
            "title": "Has Password",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "alias",
          "eligibility",
          "enabled",
          "has_password"
        ],
        "title": "LearnerPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Shepherd Academy Universe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/learners": {
      "get": {
        "operationId": "learners_api_v1_admin_learners_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearnerPublic"
                  },
                  "title": "Response Learners Api V1 Admin Learners Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Learners",
        "tags": [
          "learners"
        ]
      },
      "post": {
        "operationId": "create_learner_api_v1_admin_learners_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Learner",
        "tags": [
          "learners"
        ]
      }
    },
    "/api/v1/admin/learners/{learner_id}": {
      "delete": {
        "operationId": "delete_learner_api_v1_admin_learners__learner_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledged"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Learner",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/admin/learners/{learner_id}/account": {
      "patch": {
        "operationId": "update_account_api_v1_admin_learners__learner_id__account_patch",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerAccountUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Account",
        "tags": [
          "learners"
        ]
      }
    },
    "/api/v1/admin/learners/{learner_id}/devices": {
      "get": {
        "operationId": "learner_devices_api_v1_admin_learners__learner_id__devices_get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearnerDevicePublic"
                  },
                  "title": "Response Learner Devices Api V1 Admin Learners  Learner Id  Devices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Learner Devices",
        "tags": [
          "learners"
        ]
      }
    },
    "/api/v1/admin/learners/{learner_id}/devices/{device_id}": {
      "delete": {
        "operationId": "revoke_device_api_v1_admin_learners__learner_id__devices__device_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learner Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledged"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Device",
        "tags": [
          "learners"
        ]
      }
    },
...
    "/api/v1/admin/learners/{learner_id}/revoke": {
      "post": {
        "operationId": "revoke_devices_api_v1_admin_learners__learner_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledged"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Devices",
        "tags": [
          "learners"
        ]
      }
    },
...
  }
}